+++ netbox_rir_manager/__init__.py
@@ -2,7 +2,7 @@ import logging
 
 from netbox.plugins import PluginConfig
 
-__version__ = "0.3.2"
+__version__ = "0.4.0"
 
 logger = logging.getLogger(__name__)